Fixed style deletion
