+++ e2e/tests/static/doc/cookies/default.fqlx
