+++ e2e/tests/static/doc/cookies/get.fqlx
